Calculate the kernel-matrix for a pathway

Uses individuals' genotypes to create a kernel object including the calculated kernel matrix for a specific pathway. Each numeric value within this matrix is calculated from two individuals' genotypevectors of the SNPs within the pathway by a kernel function. It can be interpreted as the genetic similiarity of the individuals. Association between the pathway and a binary phenotype (case-control status) can be evaluated in the logistic kernel machine test, based on the kernel object. Three kernel functions are available.

# S4 method for GWASdata
calc_kernel(
  object,
  pathway,
  knots = NULL,
  type = c("lin", "sia", "net"),
  calculation = c("cpu", "gpu"),
  ...
)

# S4 method for GWASdata
lin_kernel(object, pathway, knots = NULL, calculation = c("cpu", "gpu"), ...)

# S4 method for GWASdata
sia_kernel(object, pathway, knots = NULL, calculation = c("cpu", "gpu"), ...)

# S4 method for GWASdata
net_kernel(object, pathway, knots = NULL, calculation = c("cpu", "gpu"), ...)

Arguments

object

GWASdata object containing the genotypes of the individuals for which a kernel will be calculated.

pathway

object of the class pathway specifying the SNP set for which a kernel will be calculated.

knots

GWASdata object, if specified a kernel will be computed.

type

character indicating the kernel type: Use 'lin' to specify the linear kernel, 'sia' for the size-adjusted or 'net' for the network-based kernel.

calculation

character specifying if the kernel matrix is computed on CPU or GPU.

...

further arguments to be passed to kernel computations.

Value

Returns an object of class kernel, including the similarity matrix of the pathway for the considered individuals.

If knots are specified low-rank kernel of class a lowrank_kernel

will be returned, which is not necessarily quadratic and symmetric.

Details

Different types of kernels can be constructed:

type='lin' creates the linear kernel assuming additive SNP effects to be evaluated in the logistic kernel machine test.

type='sia' calculates the size-adjusted kernel which takes into consideration the numbers of SNPs and genes in a pathway to correct for size bias.

type='net' calculates the network-based kernel. Here not only information on gene membership and gene/pathway size in number of SNPs is incorporated, but also the interaction structure of genes in the pathway.

For more details, check the references.

Methods (by class)

lin_kernel(GWASdata):

sia_kernel(GWASdata):

net_kernel(GWASdata):

References

Wu MC, Kraft P, Epstein MP, Taylor DM, Chanock SJ, Hunter DJ, Lin X Powerful SNP-Set Analysis for Case-Control Genome-Wide Association Studies. Am J Hum Genet 2010, 86:929-42

Freytag S, Bickeboeller H, Amos CI, Kneib T, Schlather M: A Novel Kernel for Correcting Size Bias in the Logistic Kernel Machine Test with an Application to Rheumatoid Arthritis. Hum Hered. 2012, 74(2):97-108.

Freytag S, Manitz J, Schlather M, Kneib T, Amos CI, Risch A, Chang-Claude J, Heinrich J, Bickeboeller H: A network-based kernel machine test for the identification of risk pathways in genome-wide association studies. Hum Hered. 2013, 76(2):64-75.

See also

Author

Stefanie Friedrichs, Juliane Manitz

Examples

data(gwas)
data(hsa04020)
lin_kernel <- calc_kernel(gwas, hsa04020, knots=NULL, type='lin', calculation='cpu')
summary(lin_kernel)
#> An object of class kernel of type lin for pathway hsa04020 with values: 
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    1208    1315    1349    1367    1383    2395 
#> 
#> Number of Individuals: 50 
#> The pathway contains 180 genes.
net_kernel <- calc_kernel(gwas, hsa04020, knots=NULL, type='net', calculation='cpu')
summary(net_kernel)
#> An object of class kernel of type network for pathway hsa04020 with values: 
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    1536    1636    1676    1679    1718    1981 
#> 
#> Number of Individuals: 50 
#> The pathway contains 180 genes.