An S4 class to represent the variance component test.

An S4 class to represent the variance component test.

show displays basic information on lkmt object

summary generates a lkmt object summary including the used kernel, pathway and the test result

# S4 method for lkmt
show(object)

# S4 method for lkmt
summary(object)

Arguments

object

An object of class lkmt.

Value

show Basic information on lkmt object.

summary Summarized information on lkmt object.

Slots

formula

A formula stating the regression nullmodel that will be used in the variance component test.

kernel

An object of class kernel representing the similarity matrix of the individuals based on which the pathways influence is evaluated.

GWASdata

An object of class GWASdata including the data on which the test is conducted.

statistic

A vector giving the value of the variance component test statistic.

df

A vector containing the number of degrees of freedom.

p.value

A vector giving the p-value calculated for the pathway object considered in the variance component test.

For details on the variance component test see the references.

References

Liu D, Lin X, Ghosh D: Semiparametric regression of multidimensional genetic pathway data: least-squares kernel machines and linear mixed models. Biometrics 2007, 63(4):1079-88.

Wu MC, Kraft P, Epstein MP, Taylor DM, Chanock SJ, Hunter DJ, Lin X: Powerful SNP-Set Analysis for Case-Control Genome-Wide Association Studies. Am J Hum Genet 2010, 86:929-42

Author

Juliane Manitz, Stefanie Friedrichs

Examples

data(hsa04020)
data(gwas)
# compute kernel
net_kernel <- calc_kernel(gwas, hsa04020, knots=NULL, type='net', calculation='cpu')
# perform LKMT 
res <- lkmt_test(pheno ~ sex + age, net_kernel, gwas, method='satt')
# show and summary methods
show(res)
#> An object of class lkmt:
#> GWAS data: example study 
#> Kernel: network 
#> Pathway: hsa04020 
#> 
#> The score test results in the p-value:  0.3711717 
#> 
summary(res)
#> An object of class lkmt:
#> 
#> GWAS data: example study 
#> Kernel: network 
#> Pathway: hsa04020 
#> 
#> 	Dan Schaid Score Test
#> 
#> Chi Squared Test Statictic Value = 38.63032 
#> Degrees of Freedom = 36.44564 
#> p-value = 0.3711717 
#> 
# summary method
summary(lkmt.net.kernel.hsa04020)
#> An object of class lkmt:
#> 
#> GWAS data: example study 
#> Kernel: network 
#> Pathway: hsa04020 
#> 
#> 	Dan Schaid Score Test
#> 
#> Chi Squared Test Statictic Value = 38.63032 
#> Degrees of Freedom = 36.44564 
#> p-value = 0.3711717 
#>