Example test result for the network-based kernel for pathway hsa04020.

An object of class lkmt containing exemplary test results for an application of the logistic kernel machine test, derived from the example data.

data(lkmt.net.kernel.hsa04020)

Format

An object of class lkmt for the network-based kernel and the pathway hsa04020.

formular

gives a formular defining the nullmodel used in the logistic kernel machine test

kernel

includes the kernel object of the pathway to be evaluated

GWASdata

gives the GWASdata object including the study data considered in testing

statistic

gives the value of the test statistic

df

specifies the degrees of freedom

p.value

includes teh p-value resulting from the test

Source

simulated data and Ensembl extract

Examples

data(hsa04020)
data(gwas)
net_kernel <- calc_kernel(gwas, hsa04020, knots=NULL, type='net', calculation='cpu')
lkmt_test(pheno ~ sex + age, net_kernel, gwas, method='satt')
#> An object of class lkmt:
#> GWAS data: example study 
#> Kernel: network 
#> Pathway: hsa04020 
#> 
#> The score test results in the p-value:  0.3711717 
#>