A function to calculate the p-values for kernel matrices.

For parameter 'satt' a pathway's influence on the probability of beeing a case is evaluated in the logistic kernel machine test and p-values are determined using a Sattherthwaite approximation as described by Dan Schaid.

For parameter 'davies' a pathways influence on the probability of beeing a case is evaluated using the p-value calculation method described by Davies. Here the function davies from package CompQuadForm is used.

lkmt_test(formula, kernel, GWASdata, method = c("satt", "davies"), ...)

# S4 method for matrix
score_test(x1, x2)

# S4 method for matrix
davies_test(x1, x2)

Arguments

formula

The formula to be used for the regression nullmodel.

kernel

An object of class kernel including the pathway representing kernel-matrix based on which the test statistic will be calculated.

GWASdata

A GWASdata object stating the data used in analysis.

method

A character specifying which method will be used for p-value calculation. Available are 'satt' for the Satterthwaite approximation and 'davies' for Davies' algorithm. For more details see the references.

...

Further arguments can be given to the function.

x1

A matrix which is the similarity matrix calculated for the pathway to be tested.

x2

An lm or glm object of the nullmodel with fixed effects covariates included, but no genetic random effects.

Value

An lkmt object including the following test results

The formula of the regression nullmodel used in the variance component test.

An object of class kernel including the similarity matrix of the individuals based on which the pathways influence is evaluated.

An object of class GWASdata stating the data on which the test was conducted.

statistic A vector giving the value of the variance component test statistic.

df A vector giving the number of degrees of freedom.

p.value A vector giving the p-value calculated for the pathway in the variance component test.

References

For details on the variance component test

Wu MC, Kraft P, Epstein MP, Taylor DM, Chanock SJ, Hunter DJ, Lin X: Powerful SNP-Set Analysis for Case-Control Genome-Wide Association Studies. Am J Hum Genet 2010, 86:929-42

Liu D, Lin X, Ghosh D: Semiparametric regression of multidimensional genetic pathway data: least-squares kernel machines and linear mixed models. Biometrics 2007, 63(4):1079-88.

For details on the p-value calculation see

Schaid DJ: Genomic Similarity and Kernel Methods I: Advancements by Building on Mathematical and Statistical Foundations. Hum Hered 2010, 70:109-31

Davies R: Algorithm as 155: the distribution of a linear combination of chi-2 random variables. J R Stat Soc Ser C 1980, 29:323-333.

Author

Stefanie Friedrichs, Juliane Manitz

Examples

data(hsa04020)
data(gwas)
net_kernel <- calc_kernel(gwas, hsa04020, knots=NULL, type='net', calculation='cpu')
lkmt_test(pheno ~ sex + age, net_kernel, gwas, method='satt')
#> An object of class lkmt:
#> GWAS data: example study 
#> Kernel: network 
#> Pathway: hsa04020 
#> 
#> The score test results in the p-value:  0.3711717 
#>