An S4 class for an object assigning genes to pathways

This function lists all genes formig a particular pathway. Start and end positions of these genes are extracted from the Ensemble database. The database is accessed via the R-package biomaRt.

pathway_info(x)

# S4 method for character
pathway_info(x)

# S4 method for pathway_info
show(object)

# S4 method for pathway_info
summary(object)

Arguments

x

A character identifying the pathway for which gene infomation should be extracted. Here KEGG IDs (format: 'hsa00100') are used.

object

An object of class pathway_info.

Value

A data.frame including as many rows as genes appear in the pathway. for each gene its name, the start and end point and the chromosome it lies on are given.

show Basic information on pathway_info object.

summary Summarized information on pathway_info object.

Slots

info

A data.frame including information on genes contained in pathways with columns 'pathway', 'gene_start', 'gene_end', 'chr' and 'gene'.

See also

Author

Stefanie Friedrichs, Juliane Manitz

Examples

data(hsa04022_info) # pathway_info('hsa04020') 
show(hsa04022_info)
#> An object of class pathway_info
#> Number of pathways: 1 
#> Number of genes: 161 
#> First six rows: 
#>    pathway gene_start  gene_end chr  gene
#> 1 hsa04022   45574140  45723116   7 ADCY1
#> 2 hsa04022    7396208   7830081   5 ADCY2
#> 3 hsa04022   24819169  24919839   2 ADCY3
#> 4 hsa04022   24318349  24335093  14 ADCY4
#> 5 hsa04022  123282296 123449758   3 ADCY5
#> 6 hsa04022   48766194  48789037  12 ADCY6
summary(hsa04022_info)
#> An object of class pathway_info
#> Number of pathways: 1 
#> Number of genes: 161 
#> First six rows: 
#>    pathway gene_start  gene_end chr  gene
#> 1 hsa04022   45574140  45723116   7 ADCY1
#> 2 hsa04022    7396208   7830081   5 ADCY2
#> 3 hsa04022   24819169  24919839   2 ADCY3
#> 4 hsa04022   24318349  24335093  14 ADCY4
#> 5 hsa04022  123282296 123449758   3 ADCY5
#> 6 hsa04022   48766194  48789037  12 ADCY6