read genotype data from file to one of several available objects, which can be passed to a GWASdata object GWASdata.

read genotype data from file to one of several available objects, which can be passed to a GWASdata object GWASdata.

# S4 method for character
read_geno(
  file.path,
  save.path = NULL,
  sep = " ",
  header = TRUE,
  use.fread = TRUE,
  use.big = FALSE,
  row.names = FALSE,
  ...
)

Arguments

file.path

character giving the path to the data file to be read

save.path

character containing the path for the backingfile

sep

character. A field delimeter. See read.big.matrix for details.

header

logical. Does the data set contain column names?

use.fread

logical. Should the dataset be read using the function fread fread from package data.table?

use.big

logical. Should the dataset be read using the function read.big.matrix from package bigmemory?

row.names

logical. Does the dataset include rownames?

...

further arguments to be passed to read_geno.

Details

If the data set contains rownames specified, set option has.row.names = TRUE.

Examples

if (FALSE) {
path <- system.file("extdata", "geno.txt", package = "kangar00")
geno <- read_geno(path, save.path = getwd(), sep = " ", use.fread = FALSE, row.names = FALSE)
}